Fix cors

+++ Catherine.Web/src/app/country/country-list/country-list.component.ts
@@ -25,6 +25,7 @@ export class CountryListComponent implements OnInit {
   }
 
   onDelete(id) {
+    if(!confirm('Jeste li sigurni?')) return;
     this.countryService
       .deleteOne(id)
       .subscribe(_ => this.getCountries())
